Remove target framework moniker parsing code
This change removes the calls to the IVsFrameworkParser API. This makes it possible to differentiate between TFMs with and without OS versions. It also addresses #6674 by displaying the full TFM in the dependencies node.

The change also replaces some references to ConfigurationGeneral.TargetFramework with ConfigurationGeneral.TargetFrameworkMoniker since the latter is the correct way to get the TFM for projects that are not cross-targeting (see AggregateCrossTargetProjectContextProvider and DependenciesSnapshotProvider.ContextTracker)